fix: fallback colour handling in __jolly_make_sep_link
In case of the first two calls to lp_terminal_format when bg is -1
fallback can't be used anyway.  With the third call it is assumed that
the glyphs' foreground form the separator's shape and the bacground color
passed to __jolly_make_sep_link is used as the foreground color. The
same needs to be done for the fallback colour.