Introduce JOLLY_COLOR_HOSTNAME
It is now possible to configure colors of the hostname segment independently.
Although it may seem contrary to the the spirit of the Jolly theme to limit
distractions for the user, it may actually help users quickly distingush
hosts they are working with.

If not set in a configuration file JOLLY_COLOR_HOSTNAME is initialized
with the value of JOLLY_COLOR_LITE.